Move `NormalizeIP` to netutil
