+++ go/netutil/netutil.go
@@ -189,3 +189,17 @@ func ResolveIPv4Addrs(addr string) ([]string, error) {
 	}
 	return result, nil
 }
+
+// NormalizeIP normalizes loopback addresses to avoid spurious errors when
+// communicating to different representations of the loopback.
+//
+// Note: this also maps IPv6 localhost to IPv4 localhost, as
+// TabletManagerClient.GetReplicas() (the only place this function is used on)
+// will return only IPv4 addresses.
+func NormalizeIP(s string) string {
+	if ip := net.ParseIP(s); ip != nil && ip.IsLoopback() {
+		return "127.0.0.1"
+	}
+
+	return s
+}
