+++ go/netutil/netutil_test.go
@@ -171,3 +171,20 @@ func TestResolveIPv4Addrs(t *testing.T) {
 		})
 	}
 }
+
+func TestNormalizeIP(t *testing.T) {
+	table := map[string]string{
+		"1.2.3.4":   "1.2.3.4",
+		"127.0.0.1": "127.0.0.1",
+		"127.0.1.1": "127.0.0.1",
+		// IPv6 must be mapped to IPv4.
+		"::1": "127.0.0.1",
+		// An unparseable IP should be returned as is.
+		"127.": "127.",
+	}
+	for input, want := range table {
+		if got := NormalizeIP(input); got != want {
+			t.Errorf("NormalizeIP(%#v) = %#v, want %#v", input, got, want)
+		}
+	}
+}
